Input: elantech - Remove the ETP_SAMPLE_QUERY read that is now redundant
commit 6b30c73e9f37183ad60c7f7050acf8e8edf91e9c introduced a read of
ETP_SAMPLE_QUERY to etd->samples on line 1662. Because of that
introduction, this other read of ETP_SAMPLE_QUERY at line 1040 becomes
redundant and the already read etd->samples can be used instead.

This line 1040 read will only be called when
(etd->fw_version >> 16) == 0x14 and consequentially
hardware version != 1 so the line 1662 read is also called.

Signed-off-by: Ulrik De Bie <ulrik.debie-os@e2big.org>
CC: Duson Lin <dusonlin@emc.com.tw>